[fix] display the proper duration in minutes w/o TZ offset
The date function uses the local timezone by default. But $minutes has no timezone info, hence the hack with ‘- 60’ was introduced to fix the display in some configuration. Depending on config and with recent evolution, this is counter-intuitive and produces some edge case (for exemple, a duration of 0 would then display a duration of 23h00).

Use gmdate instead, at least we always have a “fake” TZ offset of 0.