🐛(migration) fix migrations script
From scratch it wasn't possible to install the application. The
migration crashed at Version20190218130524_job_id_not_null step. The
usage of the entity manager and querying the entity use the last
available schema and at this step some colum where not present. To avoid
this entity manager should not be used in migrations script.